+++ Content.Server/Store/Systems/StoreSystem.cs
@@ -1,39 +1,23 @@
-using System.Linq;
-using Content.Server.Store.Components;
-using Content.Shared.FixedPoint;
 using Content.Shared.Implants.Components;
-using Content.Shared.Interaction;
-using Content.Shared.Popups;
-using Content.Shared.Stacks;
 using Content.Shared.Store;
 using Content.Shared.Store.Components;
-using Content.Shared.Store.Events;
 using Content.Shared.UserInterface;
-using Robust.Shared.Prototypes;
-using Robust.Shared.Timing;
 using Robust.Shared.Utility;
 
 namespace Content.Server.Store.Systems;
 
 public sealed partial class StoreSystem : SharedStoreSystem
 {
-    [Dependency] private readonly IPrototypeManager _proto = default!;
-    [Dependency] private readonly SharedPopupSystem _popup = default!;
-    [Dependency] private readonly IGameTiming _timing = default!;
-    [Dependency] private readonly SharedUserInterfaceSystem _uiSystem = default!;
-
     public override void Initialize()
     {
         base.Initialize();
 
         SubscribeLocalEvent<StoreComponent, ActivatableUIOpenAttemptEvent>(OnStoreOpenAttempt);
-        SubscribeLocalEvent<CurrencyComponent, AfterInteractEvent>(OnAfterInteract);
         SubscribeLocalEvent<StoreComponent, BeforeActivatableUIOpenEvent>(BeforeActivatableUiOpen);
 
         SubscribeLocalEvent<StoreComponent, MapInitEvent>(OnMapInit);
         SubscribeLocalEvent<StoreComponent, ComponentStartup>(OnStartup);
         SubscribeLocalEvent<StoreComponent, ComponentShutdown>(OnShutdown);
-        SubscribeLocalEvent<StoreComponent, IntrinsicStoreActionEvent>(OnIntrinsicStoreAction);
 
         SubscribeLocalEvent<RemoteStoreComponent, OpenUplinkImplantEvent>(OnImplantActivate);
 
@@ -48,8 +32,8 @@ public sealed partial class StoreSystem : SharedStoreSystem
         component.StartingMap = Transform(uid).MapUid;
 
         // Add the bui key if it does not exist already (the check is needed to make sure that we don't overwrite existing InterfaceData).
-        if (!_uiSystem.HasUi(uid, StoreUiKey.Key))
-            _uiSystem.SetUi(uid, StoreUiKey.Key, new InterfaceData("StoreBoundUserInterface"));
+        if (!UI.HasUi(uid, StoreUiKey.Key))
+            UI.SetUi(uid, StoreUiKey.Key, new InterfaceData("StoreBoundUserInterface"));
     }
 
     private void OnStartup(EntityUid uid, StoreComponent component, ComponentStartup args)
@@ -75,7 +59,7 @@ public sealed partial class StoreSystem : SharedStoreSystem
         if (!component.OwnerOnly)
             return;
 
-        if (!_mind.TryGetMind(args.User, out var mind, out _))
+        if (!Mind.TryGetMind(args.User, out var mind, out _))
             return;
 
         component.AccountOwner ??= mind;
@@ -85,30 +69,9 @@ public sealed partial class StoreSystem : SharedStoreSystem
             return;
 
         if (!args.Silent)
-            _popup.PopupEntity(Loc.GetString("store-not-account-owner", ("store", uid)), uid, args.User);
+            Popup.PopupEntity(Loc.GetString("store-not-account-owner", ("store", uid)), uid, args.User);
 
         args.Cancel();
-    }
-
-    private void OnAfterInteract(EntityUid uid, CurrencyComponent component, AfterInteractEvent args)
-    {
-        if (args.Handled || !args.CanReach || args.Target is not { } target)
-            return;
-
-        if (!TryGetStore(target, out var store))
-            return;
-
-        var ev = new CurrencyInsertAttemptEvent(args.User, target, args.Used, store.Value.Comp);
-        RaiseLocalEvent(target, ev);
-        if (ev.Cancelled)
-            return;
-
-        if (!TryAddCurrency((uid, component), (store.Value, store.Value.Comp)))
-            return;
-
-        args.Handled = true;
-        var msg = Loc.GetString("store-currency-inserted", ("used", args.Used), ("target", target));
-        _popup.PopupEntity(msg, target, args.User);
     }
 
     private void OnImplantActivate(Entity<RemoteStoreComponent> entity, ref OpenUplinkImplantEvent args)
@@ -117,104 +80,5 @@ public sealed partial class StoreSystem : SharedStoreSystem
             return;
 
         ToggleUi(args.Performer, store, store.Comp, entity, entity.Comp);
-    }
-
-    /// <summary>
-    /// Gets the value from an entity's currency component.
-    /// Scales with stacks.
-    /// </summary>
-    /// <remarks>
-    /// If this result is intended to be used with <see cref="TryAddCurrency(Robust.Shared.GameObjects.Entity{Content.Server.Store.Components.CurrencyComponent?},Robust.Shared.GameObjects.Entity{Content.Shared.Store.Components.StoreComponent?})"/>,
-    /// consider using <see cref="TryAddCurrency(Robust.Shared.GameObjects.Entity{Content.Server.Store.Components.CurrencyComponent?},Robust.Shared.GameObjects.Entity{Content.Shared.Store.Components.StoreComponent?})"/> instead to ensure that the currency is consumed in the process.
-    /// </remarks>
-    /// <param name="uid"></param>
-    /// <param name="component"></param>
-    /// <returns>The value of the currency</returns>
-    public Dictionary<string, FixedPoint2> GetCurrencyValue(EntityUid uid, CurrencyComponent component)
-    {
-        var amount = EntityManager.GetComponentOrNull<StackComponent>(uid)?.Count ?? 1;
-        return component.Price.ToDictionary(v => v.Key, p => p.Value * amount);
-    }
-
-    /// <summary>
-    /// Tries to add a currency to a store's balance. Note that if successful, this will consume the currency in the process.
-    /// </summary>
-    public bool TryAddCurrency(Entity<CurrencyComponent?> currency, Entity<StoreComponent?> store)
-    {
-        if (!Resolve(currency.Owner, ref currency.Comp))
-            return false;
-
-        if (!Resolve(store.Owner, ref store.Comp))
-            return false;
-
-        var value = currency.Comp.Price;
-        if (TryComp(currency.Owner, out StackComponent? stack) && stack.Count != 1)
-        {
-            value = currency.Comp.Price
-                .ToDictionary(v => v.Key, p => p.Value * stack.Count);
-        }
-
-        if (!TryAddCurrency(value, store, store.Comp))
-            return false;
-
-        // Avoid having the currency accidentally be re-used. E.g., if multiple clients try to use the currency in the
-        // same tick
-        currency.Comp.Price.Clear();
-        if (stack != null)
-            _stack.SetCount((currency.Owner, stack), 0);
-
-        QueueDel(currency);
-        return true;
-    }
-
-    /// <summary>
-    /// Tries to add a currency to a store's balance
-    /// </summary>
-    /// <param name="currency">The value to add to the store</param>
-    /// <param name="uid"></param>
-    /// <param name="store">The store to add it to</param>
-    /// <returns>Whether or not the currency was succesfully added</returns>
-    public bool TryAddCurrency(Dictionary<string, FixedPoint2> currency, EntityUid uid, StoreComponent? store = null)
-    {
-        if (!Resolve(uid, ref store))
-            return false;
-
-        //verify these before values are modified
-        foreach (var type in currency)
-        {
-            if (!store.CurrencyWhitelist.Contains(type.Key))
-                return false;
-        }
-
-        foreach (var type in currency)
-        {
-            if (!store.Balance.TryAdd(type.Key, type.Value))
-                store.Balance[type.Key] += type.Value;
-        }
-
-        UpdateUserInterface(null, uid, store);
-        return true;
-    }
-
-    private void OnIntrinsicStoreAction(Entity<StoreComponent> ent, ref IntrinsicStoreActionEvent args)
-    {
-        ToggleUi(args.Performer, ent.Owner, ent.Comp);
-    }
-
-}
-
-public sealed class CurrencyInsertAttemptEvent : CancellableEntityEventArgs
-{
-    public readonly EntityUid User;
-    public readonly EntityUid Target;
-    public readonly EntityUid Used;
-    public readonly StoreComponent Store;
-
-    public CurrencyInsertAttemptEvent(EntityUid user, EntityUid target, EntityUid used, StoreComponent store)
-    {
-        User = user;
-        Target = target;
-        Used = used;
-        Store = store;
     }
 }
